requirements.txt
keyboard>=0.13.0
apscheduler>=3.10.0
plyer>=2.1.0
python-dateutil>=2.8.0
win10toast>=0.9